notifications: typescript refactor
