+++ src/components/Notification.svelte
@@ -1,6 +1,6 @@
 <script>
 
-  import { notification, NotificationType } from "../stores/notificationStore.js";
+  import { notification } from "../stores/notificationStore.ts";
   import { cubicIn } from "svelte/easing";
   import { fly, scale, fade } from "svelte/transition";
   import errorIcon from "../assets/notification/error.svg";
@@ -13,7 +13,7 @@
     timeout = null,
     hover = false;
 
-  $: notificationClass = $notification === null ? "" : Object.keys(NotificationType).find(key => NotificationType[key] === $notification.type);
+  $: notificationClass = $notification === null ? "" : $notification.type;
 
   function hide () {
     notification.remove();
