+++ src/routes/login.svelte
@@ -5,7 +5,8 @@
   import stroke from "../assets/stroke.svg";
   import { fade } from "svelte/transition";
   import { keyfile, loggedIn, address } from "../stores/keyfileStore.js";
-  import { notification, NotificationType } from "../stores/notificationStore.js";
+  import { NotificationType } from "../types.ts";
+  import { notification } from "../stores/notificationStore.ts";
   import Arweave from "arweave";
 
   let isDragOver = false;
