+++ src/stores/notificationStore.ts
@@ -1,4 +1,5 @@
 import { writable } from "svelte/store";
+import type { NotificationType } from "../types";
 
 export const notification = createNotificationStore();
 
@@ -7,18 +8,10 @@ function createNotificationStore () {
   set(null);
 
   return {
-    notify (title, description, type, timeout) {
-      if(typeof type !== "number" || type > 3) return;
+    notify (title: string, description: string, type: NotificationType, timeout: number) {
       set({ title, description, type, timeout });
     },
     subscribe,
     remove: () => set(null)
   }
-}
-
-export const NotificationType = {
-  error: 0,
-  warning: 1,
-  log: 2,
-  success: 3
 }
