+++ src/types.ts
@@ -1,0 +1,6 @@
+export enum NotificationType {
+  error = "error",
+  warning = "warning",
+  log = "log",
+  success = "success"
+}
